requirements.txt
langchain-anthropic
langchain-core
flask
python-dotenv
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp

sigil-sdk
sigil-sdk-langchain
